refactor: extract markdown helpers from content-fetcher
Phase 1 of the content-fetcher refactor. Move the 9 markdown rendering
helpers from src/docs-retrieval/content-fetcher.ts to a new file
src/docs-retrieval/markdown-renderer.ts (Pattern A — Pure utility,
tier 2 per resolved decision DR-01).

Symbols moved:
- LEARNING_PATH_ICON_SVG
- EXPECT_HEADING_RE, NEXT_HEADING_RE
- escapeHtmlEntities
- inlineMarkdown
- simpleMarkdownToHtml
- wrapInOrangeOutlineList
- wrapExpectBlockInOrangeOutline
- splitAtNextHeading

The injectJourneyExtrasIntoJsonGuide function STAYS in content-fetcher.ts
(it imports generateJourneyContentWithExtras from learning-journey-helpers,
which keeps it tier 2). It now imports simpleMarkdownToHtml and
wrapExpectBlockInOrangeOutline from the new module.

Public API preserved via the docs-retrieval barrel: index.ts now
re-exports simpleMarkdownToHtml from './markdown-renderer'. The
test-only consumer (content-fetcher.test.ts) and the pre-extraction
test were updated to import directly from './markdown-renderer' — this
is an ownership-path update permitted by the wiki for Pattern A/J
("update source-owned contract tests in the same commit if needed").
No assertions were modified.

content-fetcher.ts also drops the now-unused sanitizeHtmlUrl import
(it only flowed through inlineMarkdown, which moved with the markdown
cluster).

All 15 pre-extraction assertions still pass against the new module.
content-fetcher.test.ts (24 it) still passes. typecheck clean.

Refs: .refactor/PLAN.md Phase 1, DR-01 (tier 2 destination), DR-04
(barrel re-exports), DR-07 (temporary test exports — now permanent).
